Refactoring TechSheetForm
